refactor: SearchResultRowViewをUIモジュールに移動してKingfisher依存を整理
- BookSearchResultsView.swiftを新規作成してUI層に検索結果表示機能を移動
- BookAutoFillContainerButtonから重複するSearchResultRowView実装を削除
- AdminモジュールのKingfisher直接依存を除去し、UI層でのみ使用する構成に変更
- Container/Presentationパターンを維持しながらアーキテクチャを改善
