+++ PictureBookLendingAdminApp/PictureBookLendingAdmin/Views/Book/BookAutoFillContainerButton.swift
@@ -1,4 +1,3 @@
-import Kingfisher
 import PictureBookLendingDomain
 import PictureBookLendingInfrastructure
 import PictureBookLendingModel
@@ -54,39 +53,14 @@ struct BookAutoFillContainerButton: View {
     
     @ViewBuilder
     private var searchResultsSheet: some View {
-        NavigationStack {
-            Group {
-                if registerModel.searchResults.isEmpty {
-                    ContentUnavailableView(
-                        "検索結果なし",
-                        systemImage: "magnifyingglass",
-                        description: Text("該当する絵本が見つかりませんでした")
-                    )
-                } else {
-                    List(registerModel.searchResults, id: \.book.id) { scoredBook in
-                        Button {
-                            selectBook(scoredBook.book)
-                        } label: {
-                            SearchResultRowView(scoredBook: scoredBook)
-                        }
-                        .buttonStyle(.plain)
-                    }
-                    .listStyle(.plain)
-                }
+        BookSearchResultsView(
+            searchResults: registerModel.searchResults,
+            onBookSelect: selectBook,
+            onCancel: {
+                isResultSheetPresented = false
+                registerModel.clearSearchResults()
             }
-            .navigationTitle("検索結果")
-            #if os(iOS)
-                .navigationBarTitleDisplayMode(.inline)
-            #endif
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("キャンセル") {
-                        isResultSheetPresented = false
-                        registerModel.clearSearchResults()
-                    }
-                }
-            }
-        }
+        )
     }
     
     // MARK: - Action Handlers
@@ -127,73 +101,6 @@ struct BookAutoFillContainerButton: View {
         // UI状態をリセット
         isResultSheetPresented = false
         registerModel.clearSearchResults()
-    }
-}
-
-/// 検索結果行ビュー
-private struct SearchResultRowView: View {
-    let scoredBook: ScoredBook
-    
-    var body: some View {
-        HStack {
-            // サムネイル画像
-            KFImage(URL(string: scoredBook.book.thumbnail ?? scoredBook.book.smallThumbnail ?? ""))
-                .placeholder {
-                    Image(systemName: "book.closed")
-                        .foregroundStyle(.secondary)
-                        .font(.title2)
-                }
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 60, height: 80)
-                .background(.regularMaterial)
-                .clipShape(RoundedRectangle(cornerRadius: 6))
-            
-            VStack(alignment: .leading, spacing: 4) {
-                Text(scoredBook.book.title)
-                    .font(.headline)
-                    .foregroundStyle(.primary)
-                    .multilineTextAlignment(.leading)
-                
-                Text(scoredBook.book.author)
-                    .font(.subheadline)
-                    .foregroundStyle(.secondary)
-                
-                if let publisher = scoredBook.book.publisher {
-                    Text(publisher)
-                        .font(.caption)
-                        .foregroundStyle(.tertiary)
-                }
-            }
-            
-            Spacer()
-            
-            VStack(alignment: .trailing, spacing: 4) {
-                // スコア表示
-                HStack {
-                    Text("関連度:")
-                        .font(.caption2)
-                        .foregroundStyle(.secondary)
-                    
-                    Text("\(Int(scoredBook.score * 100))%")
-                        .font(.caption)
-                        .foregroundStyle(scoreColor)
-                        .fontWeight(.medium)
-                }
-            }
-        }
-        .padding(.vertical, 4)
-    }
-    
-    private var scoreColor: Color {
-        switch scoredBook.score {
-        case 0.8...:
-            return .green
-        case 0.5..<0.8:
-            return .orange
-        default:
-            return .red
-        }
     }
 }
